Refactored again. Now the mapping of atom-type to radius is a separate function, that can be customized.
Before the exclusion of some atoms was completely automatic because only one definition of atom types was allowed, now it is less straightforward. For now hydrogens are excluded when a whole PDB-file is read, but when they are added manually one by one, it is assumed that users know what they're doing.