Add prefix to variable names in classifier_*.c
Modify config2c.pl to add those prefixes and make sure class names
are printed in the correct order (only worked by chance before?).
The ProtOr classifier is now called freesasa_protor_classifier and
freesasa_default_classifier is now a macro defining it as the ProtOr
classifier. This should make it more straightforward to change default
in the future.