Corrected freesasa_pdb_ishydrogen function
This function previously identified some non-hydrogen atoms as hydrogens if the second character in the atom name was 'H/D'. Example: CD,CD1,CD2 or OD1, OD2 etc. Due to this the number of non-H atoms being read were incorrect. 

A new line in the function has been added (defined in pdb.c)